+++ Stuff-Stash/client/src/App.js
@@ -2,8 +2,7 @@ import './App.css';
 import Navbar from './Pages/navbar';
 import { BrowserRouter as Redirect, Route, Switch } from 'react-router-dom';
 
-import Home from "./Pages/Home";
-import Helpme from './Pages/ListOfUsers';
+import CreateOrg from "./Pages/CreateOrg";
 import React from 'react';
 
 function App() {
@@ -13,8 +12,7 @@ function App() {
       <div className="content">
         <Navbar />
         <Switch>
-          <Route path="/list" component={Helpme} /> 
-          <Route path="/" component={Home} />
+          <Route path="/" component={CreateOrg} />
           <Redirect from='*' to='/' />
         </Switch>
       </div>
